本番環境用の設定を修正し、seeds.rbを復元。db/db配下の不要なファイルを削除

+++ db/schema.rb
@@ -16,7 +16,7 @@ ActiveRecord::Schema.define(version: 2025_07_28_064452) do
     t.string "name", null: false
     t.string "record_type", null: false
     t.integer "record_id", null: false
-    t.integer "blob_id", null: false
+    t.bigint "blob_id", null: false
     t.datetime "created_at", null: false
     t.index ["blob_id"], name: "index_active_storage_attachments_on_blob_id"
     t.index ["record_type", "record_id", "name", "blob_id"], name: "index_active_storage_attachments_uniqueness", unique: true
